notification-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-RESTX==1.1.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2
pydantic==1.10.9